+++ Radio3.0@claudiux/files/Radio3.0@claudiux/CHANGELOG.md
@@ -1,3 +1,7 @@
+### v1.14.0
+  * Improves sub-processes.
+  * Adds a button "Open the recordings folder" in YT tab (in this applet settings).
+
 ### v1.13.0~20231224
   * Fixes #5246. Display of volume level near icon and fixes scrolling problems with Cinnamon prior to 5.1.